renamed and reorganized folders
