+++ src/ReactBusinessCard/CardLayout/index.js
@@ -1,7 +1,7 @@
 import React, { PureComponent } from 'react'
 import PropTypes from 'prop-types'
-import { GithubIcon, LinkedinIcon, ResumeIcon } from './ReactBusinessCard/Icons'
-import './ReactBusinessCard/CardLayout/styles.css'
+import { GithubIcon, LinkedinIcon, ResumeIcon } from '../Icons'
+import './styles.css'
 
 const downloadResume = (uri, first, last) => {
   const noSpaces = string => string.replace(/\s+/g, '')
@@ -14,7 +14,7 @@ const downloadResume = (uri, first, last) => {
   document.body.removeChild(link)
 }
 
-export default class CardContent extends PureComponent {
+export default class CardLayout extends PureComponent {
   render() {
     const { width, height, cardInfo } = this.props
     const proportionedFontSize = Math.round(width / 70) // will have to look into this...
@@ -61,7 +61,7 @@ export default class CardContent extends PureComponent {
   }
 }
 
-CardContent.propTypes = {
+CardLayout.propTypes = {
   width: PropTypes.number.isRequired,
   height: PropTypes.number.isRequired,
   cardInfo: PropTypes.shape({
